Convert build-headers to js
