+++ sites/public/package.json
@@ -14,7 +14,7 @@
     "dev:listings": "cd ../../backend/core && yarn dev",
     "dev:server-wait": "wait-on \"http-get://localhost:${PORT:-3100}/listings\" && yarn dev",
     "dev:all": "concurrently \"yarn dev:listings\" \"yarn dev:server-wait\"",
-    "build:headers": "node --loader ts-node/esm scripts/build-headers.ts"
+    "build:headers": "node scripts/build-headers.js"
   },
   "dependencies": {
     "@bloom-housing/ui-components": "^0.3.10",
